refactor(data): Move subscriber.go from application to controller
close #3902
